aqi_testing/requirements.txt
alembic==1.4.1
appdirs==1.4.4
argh==0.26.2
async-generator==1.10
atomicwrites==1.4.0
azure-core==1.10.0
azure-storage-blob==12.7.1
backcall==0.2.0
black==19.10b0
blis==0.7.4
brotlipy==0.7.0
catalogue==2.0.1
catboost==0.24.4
certifi==2020.12.5
colorlover==0.3.0
combo==0.1.2
confuse==1.4.0
convertdate==2.1.3
cufflinks==0.17.3
cycler==0.10.0
cymem==2.0.5
Cython==0.29.14
databricks-cli==0.14.1
decorator==4.4.2
defusedxml==0.6.0
docker==4.4.1
docutils==0.16
entrypoints==0.3
ez_setup==0.9
Flask==1.1.2
funcy==1.15
future==0.18.2
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.12
graphviz==0.16
htmlmin==0.1.12
ImageHash==4.2.0
imbalanced-learn==0.7.0
iniconfig==1.1.1
ipywidgets==7.6.3
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
json5==0.9.5
jupyterlab==2.2.6
jupyterlab-widgets==1.0.0
kmodes==0.10.2
lightgbm==3.1.1
llvmlite==0.35.0
LunarCalendar==0.0.9
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
mlflow==1.13.1
mlxtend==0.18.0
msrest==0.6.21
murmurhash==1.0.5
mypy-extensions==0.4.3
networkx==2.5
nltk==3.5
numba==0.52.0
numexpr==2.7.2
oauthlib==3.1.0
pandas-profiling==2.10.0
parso==0.7.0
pathspec==0.7.0
pathtools==0.1.2
pathy==0.3.4
patsy==0.5.1
phik==0.11.0
plotly==4.14.1
pluggy==0.13.1
preshed==3.0.5
prometheus-flask-exporter==0.18.1
protobuf==3.14.0
py==1.10.0
pycaret==2.2.3
pycodestyle==2.6.0
pydantic==1.7.3
pyflakes==2.2.0
pyLDAvis==2.1.2
pynndescent==0.5.1
pyod==0.8.6
pyparsing==2.4.7
pystan==2.19.1.1
pytest==6.2.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
QDarkStyle==2.8.1
QtPy==1.9.0
querystring-parser==1.2.4
requests-oauthlib==1.3.0
retrying==1.3.3
Rtree==0.9.4
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.1
sip==4.19.13
smart-open==3.0.0
smmap==3.0.5
spacy==3.0.0
spacy-legacy==3.0.1
SQLAlchemy==1.3.23
sqlparse==0.4.1
srsly==2.4.0
statsmodels==0.12.1
suod==0.0.6
tabulate==0.8.7
tangled-up-in-unicode==0.0.6
terminado==0.9.2
testpath==0.4.4
textblob==0.15.3
thinc==8.0.1
threadpoolctl==2.1.0
typer==0.3.2
umap-learn==0.5.0
unroll==0.1.0
visions==0.6.0
waitress==1.4.4
wasabi==0.8.2
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wordcloud==1.8.1
wrapt==1.11.2
xgboost==1.3.3
yellowbrick==1.2.1
